app/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-semantic-conventions==0.57b0
